+++ src/main/java/com/team/jjan/config/OpenAPIConfig.java
@@ -13,11 +13,9 @@ public class OpenAPIConfig {
 
     @Bean
     public OpenAPI openApi() {
-        // SecuritySecheme명
         String jwtSchemeName = "jwtAuth";
-        // API 요청헤더에 인증정보 포함
         SecurityRequirement securityRequirement = new SecurityRequirement().addList(jwtSchemeName);
-        // SecuritySchemes 등록
+
         Components components = new Components()
                 .addSecuritySchemes(jwtSchemeName, new SecurityScheme()
                         .name(jwtSchemeName)
