--- src/main/java/com/team/jjan/join/dto/MailDto.java
@@ -1,14 +1,0 @@
-package com.team.jjan.join.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class MailDto {
-    private String address;
-    private String title;
-    private String message;
-}
