--- src/main/java/com/team/jjan/join/dto/PasswordRequestDto.java
@@ -1,13 +1,0 @@
-package com.team.jjan.join.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class PasswordRequestDto {
-    String email;
-    String nickname;
-}
