+++ src/main/java/com/team/jjan/join/dto/ValidationRequest.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ValidationRequest {
-    String data;
+    private String data;
 }
